synui_v115
changes:
I. There have been changes to the spectral calculation.
1. Spectral power was slightly off- it was including silent periods. This has been corrected.
2. Spectral power is calculated from FFT of the PSTH. This is tricky because:
	a. Multiple cycles are needed to accurately measure the signal and the mod frequency, so we can't use the cycle PSTH
	b. Using the uncycled PSTH and a short bin window (<2ms) results in a discrete y values (y = 0 or 1). FFT is usually used on analog signals (y taking on any value). Apparently this at least kind of still worked (to my surprise) - it was showing higher SP for channels following the pulses. However, to my limited knowledge of DSP, it's better to have more different y values. To do this, I averaged PSTHs by trains. Also, I would recommend using a longer bin size (like 1-2 ms).

Why does this imprecision exist? FFT is most accurate when your signal is analog and infinitely long in duration. Our signal is neither. We have spikes (discrete) and limited cycles. To make the signal analog, the best method I can think of is to bin the spikes as a PSTH. Using a longer bin window allows more spikes to be binned together, at the loss of temporal precision. In terms of spectral analysis, we lose the ability to detect periodic signals >250 Hz with  2ms window (which I wouldn't expect to be a problem).

II. Using the firing rate tab to see where the PSTH1 and PSTH2 Windows are...
I think it is good to look at this window when defining the PSTH1 and PSTH2 Windows because:
 1. visualise where the PSTH1 windows are relative to sound timestamps
 2. find mistakes
 3. notice any weird spiking patterns outside of the PSTH windows
 4. It is prettier now

Note that the Bin size window here is used to calculate the LDS first bin rate. Note that this does not have to be the same as the bin size of the PSTH bins. (e.g. if we want to use 2500 ms bins for LDSsn PSTHs, but measure LDS first bin rate over 1000 ms).

III.
Now saves and exports:
Is there a LDS-evoked response
-LDS first bin rate
-Average rate during the 10 seconds before LDS
-Average rate during user defined windows in the cycle PSTH

Is there a pulse-evoked response
- My method of is there a tone-evoked response (based on spectral power), marked in magenta asterisks

IV. There is an option to load tanks without copying the raw signal to memory. This should improve performance on computers with less memory.